Transform the proyecto to Senior

+++ retail/src/main/java/com/products/retail/model/ProductDetail.java
@@ -1,18 +1,19 @@
 package com.products.retail.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
 
-/**
- * Modelo que representa el detalle de un producto.
- */
+import java.math.BigDecimal;
+
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDetail {
+
     private String id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private boolean availability;
-} 
+}
